// src/api/generated/schemas/firmware.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * NervesHub API
 * NervesHub firmware update and device management API
 * OpenAPI spec version: 2.0.0
 */

export interface Firmware {
  uuid?: string;
  architecture?: string;
  platform?: string;
  version?: string;
  author?: string;
  description?: string;
  misc?: string;
  vcs_identifier?: string;
  fwup_version?: string;
  signed?: boolean;
  inserted_at?: string;
  updated_at?: string;
}